added database connection and methods to work with it

+++ app/src/main/java/com/example/coursework/screens/ExercisesScreen.kt
@@ -18,7 +18,7 @@ import com.example.coursework.composables.TopBar
 import com.example.coursework.data.TopicsRepository
 
 @Composable
-fun ExercisesScreen(navController: NavController, topicId: Int) {
+fun ExercisesScreen(navController: NavController, topicId: Long) {
     val context = LocalContext.current
     val topicsRepository = TopicsRepository(context)
     val topic = topicsRepository.findTopic(topicId)
